+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 Release 2.0.2:
- - tbd
+ - `vclib-openid`: Add response modes for query and fragment, i.e. Wallet may return the authentication response in query params or as fragment params on a SIOPv2 call
 
 Release 2.0.1:
  - `vclib-openid`: Remove `OidcSiopProtocol`, replace with `OidcSiopVerifier` and `OidcSiopWallet`, remove `AuthenticationResponse` and `AuthenticationRequest` holder classes